+++ api/openweathermap.js
@@ -57,7 +57,7 @@ var update = function (callback) {
 
     let weather = {};
     Openweathermap.getAllWeather(function (err, jsonObj) {
-        if (!err && jsonObj['coord']) {
+        if (!err && jsonObj.coord) {
             // Current weather report
             parseReport(weather, jsonObj, callback);
         }
@@ -69,7 +69,7 @@ var update = function (callback) {
     });
 
     Openweathermap.getWeatherForecast(function (err, jsonObj) {
-        if (!err && jsonObj['city']) {
+        if (!err && jsonObj.city) {
             parseForecasts(weather, jsonObj, callback);
         }
         else {
@@ -82,17 +82,17 @@ var update = function (callback) {
 
 var parseReport = function (weather, values, callback) {
     let report = weather.report || {};
-    const timezone = geoTz(values['coord']['lat'], values['coord']['lon']);
+    const timezone = geoTz(values.coord.lat, values.coord.lon);
 
-    report.AirPressure = parseInt(values['main']['pressure']);
-    report.CloudCover = parseInt(values['clouds']['all']);
-    report.Condition = values['weather'][0]['description'];
-    report.ConditionCategory = converter.getConditionCategoryOwm(values['weather'][0]['id']);
-    report.Humidity = parseInt(values['main']['humidity']);
-    report.ObservationTime = moment.unix(values['dt']).tz(timezone).format('HH:mm:ss');
-    report.Temperature = values['main']['temp'];
-    report.WindDirection = converter.getWindDirection(values['wind']['deg']);
-    report.WindSpeed = values['wind']['speed'];
+    report.AirPressure = parseInt(values.main.pressure);
+    report.CloudCover = parseInt(values.clouds.all);
+    report.Condition = values.weather[0].description;
+    report.ConditionCategory = converter.getConditionCategoryOwm(values.weather[0].id);
+    report.Humidity = parseInt(values.main.humidity);
+    report.ObservationTime = moment.unix(values.dt).tz(timezone).format('HH:mm:ss');
+    report.Temperature = values.main.temp;
+    report.WindDirection = converter.getWindDirection(values.wind.deg);
+    report.WindSpeed = values.wind.speed;
 
     weather.report = report;
     if (weather.forecasts) {
@@ -101,11 +101,11 @@ var parseReport = function (weather, values, callback) {
 };
 
 var parseForecasts = function (weather, values, callback) {
-    const timezone = geoTz(values['city']['coord']['lat'], values['city']['coord']['lon']);
+    const timezone = geoTz(values.city.coord.lat, values.city.coord.lon);
 
     let forecasts = [];
     // We get a forecast for 5 days with values each 3 hours.
-    const itemsFiltered = prepareForecasts(values['list'], timezone);
+    const itemsFiltered = prepareForecasts(values.list, timezone);
     for (let i = 0; i < itemsFiltered.length; i++) {
         forecasts[forecasts.length] = parseForecast(itemsFiltered[i], timezone);
     }
@@ -119,16 +119,16 @@ var parseForecasts = function (weather, values, callback) {
 var parseForecast = function (values, timezone) {
     let forecast = {};
 
-    forecast.AirPressure = parseInt(values['main']['pressure']);
-    forecast.CloudCover = parseInt(values['clouds']['all']);
-    forecast.Condition = values['weather'][0]['description'];
-    forecast.ConditionCategory = converter.getConditionCategoryOwm(values['weather'][0]['id']);
-    forecast.ForecastDay = moment.unix(values['dt']).tz(timezone).format('dddd');
-    forecast.Humidity = parseInt(values['main']['humidity']);
-    forecast.Temperature = values['main']['temp_max'];
-    forecast.TemperatureMin = values['main']['temp_min'];
-    forecast.WindDirection = converter.getWindDirection(values['wind']['deg']);
-    forecast.WindSpeed = values['wind']['speed'];
+    forecast.AirPressure = parseInt(values.main.pressure);
+    forecast.CloudCover = parseInt(values.clouds.all);
+    forecast.Condition = values.weather[0].description;
+    forecast.ConditionCategory = converter.getConditionCategoryOwm(values.weather[0].id);
+    forecast.ForecastDay = moment.unix(values.dt).tz(timezone).format('dddd');
+    forecast.Humidity = parseInt(values.main.humidity);
+    forecast.Temperature = values.main.temp_max;
+    forecast.TemperatureMin = values.main.temp_min;
+    forecast.WindDirection = converter.getWindDirection(values.wind.deg);
+    forecast.WindSpeed = values.wind.speed;
 
     return forecast;
 };
@@ -150,8 +150,8 @@ var prepareForecasts = function (forecasts, timezone) {
     // Find one for today
     for (let i = 0; i < forecasts.length; i++) {
         const listItem = forecasts[i];
-        if (listItem['dt'] <= endOfToday) {
-            const hourOfDay = moment.unix(listItem['dt']).tz(timezone).format("H");
+        if (listItem.dt <= endOfToday) {
+            const hourOfDay = moment.unix(listItem.dt).tz(timezone).format("H");
             if (parseInt(hourOfDay) >= 10 && parseInt(hourOfDay) <= 13) {
                 forecastsFiltered[forecastsFiltered.length] = forecasts[i];
             }
@@ -165,8 +165,8 @@ var prepareForecasts = function (forecasts, timezone) {
     // For all following days find exactly one value (best: noon)
     for (let i = 0; i < forecasts.length; i++) {
         const listItem = forecasts[i];
-        if (listItem['dt'] > endOfToday) {
-            const hourOfDay = moment.unix(listItem['dt']).tz(timezone).format("H");
+        if (listItem.dt > endOfToday) {
+            const hourOfDay = moment.unix(listItem.dt).tz(timezone).format("H");
             if (parseInt(hourOfDay) >= 10 && parseInt(hourOfDay) <= 13) {
                 forecastsFiltered[forecastsFiltered.length] = forecasts[i];
             }