+++ api/weatherunderground.js
@@ -56,13 +56,13 @@ var update = function (callback) {
     this.wunderground.conditions().forecast().request(this.location, function (error, response) {
         if (!error) {
             // Current weather report
-            weather.report = parseReport(response['current_observation']);
+            weather.report = parseReport(response.current_observation);
 
             // Forecasts for today and next 3 days
-            weather.forecasts.push(parseForecast(response['forecast']['simpleforecast']['forecastday'][0]));
-            weather.forecasts.push(parseForecast(response['forecast']['simpleforecast']['forecastday'][1]));
-            weather.forecasts.push(parseForecast(response['forecast']['simpleforecast']['forecastday'][2]));
-            weather.forecasts.push(parseForecast(response['forecast']['simpleforecast']['forecastday'][3]));
+            weather.forecasts.push(parseForecast(response.forecast.simpleforecast.forecastday[0]));
+            weather.forecasts.push(parseForecast(response.forecast.simpleforecast.forecastday[1]));
+            weather.forecasts.push(parseForecast(response.forecast.simpleforecast.forecastday[2]));
+            weather.forecasts.push(parseForecast(response.forecast.simpleforecast.forecastday[3]));
             callback(null, weather);
         }
         else {
@@ -76,21 +76,21 @@ var update = function (callback) {
 var parseReport = function (values) {
     let report = {};
 
-    report.AirPressure = parseInt(values['pressure_mb']);
-    report.Condition = values['weather'];
-    report.ConditionCategory = converter.getConditionCategory(values['icon']);
-    report.Humidity = parseInt(values['relative_humidity'].substr(0, values['relative_humidity'].length - 1));
-    report.ObservationStation = values['observation_location']['full'];
-    report.ObservationTime = values['observation_time_rfc822'].split(' ')[4];
-    report.Rain1h = isNaN(parseInt(values['precip_1hr_metric'])) ? 0 : parseInt(values['precip_1hr_metric']);
-    report.RainDay = isNaN(parseInt(values['precip_today_metric'])) ? 0 : parseInt(values['precip_today_metric']);
-    report.SolarRadiation = isNaN(parseInt(values['solarradiation'])) ? 0 : parseInt(values['solarradiation']);
-    report.Temperature = values['temp_c'];
-    report.UVIndex = isNaN(parseInt(values['UV'])) ? 0 : parseInt(values['UV']);
-    report.Visibility = isNaN(parseInt(values['visibility_km'])) ? 0 : parseInt(values['visibility_km']);
-    report.WindDirection = values['wind_dir'];
-    report.WindSpeed = parseFloat(values['wind_kph']);
-    report.WindSpeedMax = parseFloat(values['wind_gust_kph']);
+    report.AirPressure = parseInt(values.pressure_mb);
+    report.Condition = values.weather;
+    report.ConditionCategory = converter.getConditionCategory(values.icon);
+    report.Humidity = parseInt(values.relative_humidity.substr(0, values.relative_humidity.length - 1));
+    report.ObservationStation = values.observation_location.full;
+    report.ObservationTime = values.observation_time_rfc822.split(' ')[4];
+    report.Rain1h = isNaN(parseInt(values.precip_1hr_metric)) ? 0 : parseInt(values.precip_1hr_metric);
+    report.RainDay = isNaN(parseInt(values.precip_today_metric)) ? 0 : parseInt(values.precip_today_metric);
+    report.SolarRadiation = isNaN(parseInt(values.solarradiation)) ? 0 : parseInt(values.solarradiation);
+    report.Temperature = values.temp_c;
+    report.UVIndex = isNaN(parseInt(values.UV)) ? 0 : parseInt(values.UV);
+    report.Visibility = isNaN(parseInt(values.visibility_km)) ? 0 : parseInt(values.visibility_km);
+    report.WindDirection = values.wind_dir;
+    report.WindSpeed = parseFloat(values.wind_kph);
+    report.WindSpeedMax = parseFloat(values.wind_gust_kph);
 
     return report;
 };
@@ -98,17 +98,17 @@ var parseReport = function (values) {
 var parseForecast = function (values) {
     let forecast = {};
 
-    forecast.Condition = values['conditions'];
-    forecast.ConditionCategory = converter.getConditionCategory(values['icon']);
-    forecast.ForecastDay = values['date']['weekday'];
-    forecast.Humidity = parseInt(values['avehumidity'])
-    forecast.RainChance = values['pop'];
-    forecast.RainDay = isNaN(parseInt(values['qpf_allday']['mm'])) ? 0 : parseInt(values['qpf_allday']['mm']);
-    forecast.Temperature = values['high']['celsius'];
-    forecast.TemperatureMin = values['low']['celsius'];
-    forecast.WindDirection = values['avewind']['dir'];
-    forecast.WindSpeed = parseFloat(values['avewind']['kph']);
-    forecast.WindSpeedMax = parseFloat(values['maxwind']['kph']);
+    forecast.Condition = values.conditions;
+    forecast.ConditionCategory = converter.getConditionCategory(values.icon);
+    forecast.ForecastDay = values.date.weekday;
+    forecast.Humidity = parseInt(values.avehumidity);
+    forecast.RainChance = values.pop;
+    forecast.RainDay = isNaN(parseInt(values.qpf_allday.mm)) ? 0 : parseInt(values.qpf_allday.mm);
+    forecast.Temperature = values.high.celsius;
+    forecast.TemperatureMin = values.low.celsius;
+    forecast.WindDirection = values.avewind.dir;
+    forecast.WindSpeed = parseFloat(values.avewind.kph);
+    forecast.WindSpeedMax = parseFloat(values.maxwind.kph);
 
     return forecast;
 };